Revert Pack voltage to 32 bit

+++ Core/Inc/main.h
@@ -77,7 +77,7 @@ typedef struct batteryModule {
 	uint16_t cell_difference;
 	uint16_t cell_temp_lowest;
 	uint16_t cell_temp_highest;
-	uint16_t pack_voltage;
+	uint32_t pack_voltage;
 	uint16_t read_auxreg[NUM_AUXES];
 	uint16_t soc;
 	uint32_t current;
